% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgx.DMatrix.R
\name{bgx.DMatrix}
\alias{bgx.DMatrix}
\title{Construct bgx.DMatrix object}
\usage{
bgx.DMatrix(data, info = list(), missing = NA, silent = FALSE, ...)
}
\arguments{
\item{data}{a \code{matrix} object (either numeric or integer), a \code{dgCMatrix} object, or a character
string representing a filename.}

\item{info}{a named list of additional information to store in the \code{bgx.DMatrix} object.
See \code{\link{setinfo}} for the specific allowed kinds of}

\item{missing}{a float value to represents missing values in data (used only when input is a dense matrix).
It is useful when a 0 or some other extreme value represents missing values in data.}

\item{silent}{whether to suppress printing an informational message after loading from a file.}

\item{...}{the \code{info} data could be passed directly as parameters, without creating an \code{info} list.}
}
\description{
Construct bgx.DMatrix object from either a dense matrix, a sparse matrix, or a local file.
Supported input file formats are either a libsvm text file or a binary file that was created previously by
\code{\link{bgx.DMatrix.save}}).
}
\examples{
data(agaricus.train, package='tsoobgx')
train <- agaricus.train
dtrain <- bgx.DMatrix(train$data, label=train$label)
bgx.DMatrix.save(dtrain, 'bgx.DMatrix.data')
dtrain <- bgx.DMatrix('bgx.DMatrix.data')
if (file.exists('bgx.DMatrix.data')) file.remove('bgx.DMatrix.data')
}
