% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgx.DMatrix.save.R
\name{bgx.DMatrix.save}
\alias{bgx.DMatrix.save}
\title{Save bgx.DMatrix object to binary file}
\usage{
bgx.DMatrix.save(dmatrix, fname)
}
\arguments{
\item{dmatrix}{the \code{bgx.DMatrix} object}

\item{fname}{the name of the file to write.}
}
\description{
Save bgx.DMatrix object to binary file
}
\examples{
data(agaricus.train, package='tsoobgx')
train <- agaricus.train
dtrain <- bgx.DMatrix(train$data, label=train$label)
bgx.DMatrix.save(dtrain, 'bgx.DMatrix.data')
dtrain <- bgx.DMatrix('bgx.DMatrix.data')
if (file.exists('bgx.DMatrix.data')) file.remove('bgx.DMatrix.data')
}
