% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgx.dump.R
\name{bgx.dump}
\alias{bgx.dump}
\title{Dump an tsoobgx model in text format.}
\usage{
bgx.dump(model, fname = NULL, fmap = "", with_stats = FALSE,
  dump_format = c("text", "json"), ...)
}
\arguments{
\item{model}{the model object.}

\item{fname}{the name of the text file where to save the model text dump. 
If not provided or set to \code{NULL}, the model is returned as a \code{character} vector.}

\item{fmap}{feature map file representing feature types.
Detailed description could be found at 
\url{https://github.com/dmlc/tsoobgx/wiki/Binary-Classification#dump-model}.
See demo/ for walkthrough example in R, and
\url{https://github.com/dmlc/tsoobgx/blob/master/demo/data/featmap.txt} 
for example Format.}

\item{with_stats}{whether to dump some additional statistics about the splits.
When this option is on, the model dump contains two additional values:
gain is the approximate loss function gain we get in each split;
cover is the sum of second order gradient in each node.}

\item{dump_format}{either 'text' or 'json' format could be specified.}

\item{...}{currently not used}
}
\value{
If fname is not provided or set to \code{NULL} the function will return the model
as a \code{character} vector. Otherwise it will return \code{TRUE}.
}
\description{
Dump an tsoobgx model in text format.
}
\examples{
data(agaricus.train, package='tsoobgx')
data(agaricus.test, package='tsoobgx')
train <- agaricus.train
test <- agaricus.test
bst <- tsoobgx(data = train$data, label = train$label, max_depth = 2, 
               eta = 1, nthread = 2, nrounds = 2, objective = "binary:logistic")
# save the model in file 'bgx.model.dump'
dump_path = file.path(tempdir(), 'model.dump')
bgx.dump(bst, dump_path, with_stats = TRUE)

# print the model without saving it to a file
print(bgx.dump(bst, with_stats = TRUE))

# print in JSON format:
cat(bgx.dump(bst, with_stats = TRUE, dump_format='json'))

}
