% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{bgx.gblinear.history}
\alias{bgx.gblinear.history}
\title{Extract gblinear coefficients history.}
\usage{
bgx.gblinear.history(model, class_index = NULL)
}
\arguments{
\item{model}{either an \code{bgx.Booster} or a result of \code{bgx.cv()}, trained
using the \code{cb.gblinear.history()} callback.}

\item{class_index}{zero-based class index to extract the coefficients for only that
specific class in a multinomial multiclass model. When it is NULL, all the
coefficients are returned. Has no effect in non-multiclass models.}
}
\value{
For an \code{bgx.train} result, a matrix (either dense or sparse) with the columns
corresponding to iteration's coefficients (in the order as \code{bgx.dump()} would
return) and the rows corresponding to boosting iterations.

For an \code{bgx.cv} result, a list of such matrices is returned with the elements
corresponding to CV folds.
}
\description{
A helper function to extract the matrix of linear coefficients' history
from a gblinear model created while using the \code{cb.gblinear.history()}
callback.
}
