% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgx.Booster.R
\name{bgx.parameters<-}
\alias{bgx.parameters<-}
\title{Accessors for model parameters.}
\usage{
bgx.parameters(object) <- value
}
\arguments{
\item{object}{Object of class \code{bgx.Booster} or \code{bgx.Booster.handle}.}

\item{value}{a list (or an object coercible to a list) with the names of parameters to set
and the elements corresponding to parameter values.}
}
\description{
Only the setter for tsoobgx parameters is currently implemented.
}
\details{
Note that the setter would usually work more efficiently for \code{bgx.Booster.handle}
than for \code{bgx.Booster}, since only just a handle would need to be copied.
}
\examples{
data(agaricus.train, package='tsoobgx')
train <- agaricus.train

bst <- tsoobgx(data = train$data, label = train$label, max_depth = 2,
               eta = 1, nthread = 2, nrounds = 2, objective = "binary:logistic")

bgx.parameters(bst) <- list(eta = 0.1)

}
