% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgx.save.raw.R
\name{bgx.save.raw}
\alias{bgx.save.raw}
\title{Save tsoobgx model to R's raw vector,
user can call bgx.load to load the model back from raw vector}
\usage{
bgx.save.raw(model)
}
\arguments{
\item{model}{the model object.}
}
\description{
Save tsoobgx model from tsoobgx or bgx.train
}
\examples{
data(agaricus.train, package='tsoobgx')
data(agaricus.test, package='tsoobgx')
train <- agaricus.train
test <- agaricus.test
bst <- tsoobgx(data = train$data, label = train$label, max_depth = 2, 
               eta = 1, nthread = 2, nrounds = 2,objective = "binary:logistic")
raw <- bgx.save.raw(bst)
bst <- bgx.load(raw)
pred <- predict(bst, test$data)

}
