% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgx.save.R
\name{bgx.save}
\alias{bgx.save}
\title{Save tsoobgx model to binary file}
\usage{
bgx.save(model, fname)
}
\arguments{
\item{model}{model object of \code{bgx.Booster} class.}

\item{fname}{name of the file to write.}
}
\description{
Save tsoobgx model to a file in binary format.
}
\details{
This methods allows to save a model in an tsoobgx-internal binary format which is universal 
among the various tsoobgx interfaces. In R, the saved model file could be read-in later
using either the \code{\link{bgx.load}} function or the \code{bgx_model} parameter 
of \code{\link{bgx.train}}.

Note: a model can also be saved as an R-object (e.g., by using \code{\link[base]{readRDS}} 
or \code{\link[base]{save}}). However, it would then only be compatible with R, and 
corresponding R-methods would need to be used to load it.
}
\examples{
data(agaricus.train, package='tsoobgx')
data(agaricus.test, package='tsoobgx')
train <- agaricus.train
test <- agaricus.test
bst <- tsoobgx(data = train$data, label = train$label, max_depth = 2, 
               eta = 1, nthread = 2, nrounds = 2,objective = "binary:logistic")
bgx.save(bst, 'bgx.model')
bst <- bgx.load('bgx.model')
if (file.exists('bgx.model')) file.remove('bgx.model')
pred <- predict(bst, test$data)
}
\seealso{
\code{\link{bgx.load}}, \code{\link{bgx.Booster.complete}}.
}
