% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgx.DMatrix.R
\name{dim.bgx.DMatrix}
\alias{dim.bgx.DMatrix}
\title{Dimensions of bgx.DMatrix}
\usage{
\method{dim}{bgx.DMatrix}(x)
}
\arguments{
\item{x}{Object of class \code{bgx.DMatrix}}
}
\description{
Returns a vector of numbers of rows and of columns in an \code{bgx.DMatrix}.
}
\details{
Note: since \code{nrow} and \code{ncol} internally use \code{dim}, they can also
be directly used with an \code{bgx.DMatrix} object.
}
\examples{
data(agaricus.train, package='tsoobgx')
train <- agaricus.train
dtrain <- bgx.DMatrix(train$data, label=train$label)

stopifnot(nrow(dtrain) == nrow(train$data))
stopifnot(ncol(dtrain) == ncol(train$data))
stopifnot(all(dim(dtrain) == dim(train$data)))

}
