% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgx.DMatrix.R
\name{dimnames.bgx.DMatrix}
\alias{dimnames.bgx.DMatrix}
\alias{dimnames<-.bgx.DMatrix}
\title{Handling of column names of \code{bgx.DMatrix}}
\usage{
\method{dimnames}{bgx.DMatrix}(x)

\method{dimnames}{bgx.DMatrix}(x) <- value
}
\arguments{
\item{x}{object of class \code{bgx.DMatrix}}

\item{value}{a list of two elements: the first one is ignored
and the second one is column names}
}
\description{
Only column names are supported for \code{bgx.DMatrix}, thus setting of
row names would have no effect and returned row names would be NULL.
}
\details{
Generic \code{dimnames} methods are used by \code{colnames}.
Since row names are irrelevant, it is recommended to use \code{colnames} directly.
}
\examples{
data(agaricus.train, package='tsoobgx')
train <- agaricus.train
dtrain <- bgx.DMatrix(train$data, label=train$label)
dimnames(dtrain)
colnames(dtrain)
colnames(dtrain) <- make.names(1:ncol(train$data))
print(dtrain, verbose=TRUE)

}
