% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgx.DMatrix.R
\name{getinfo}
\alias{getinfo}
\alias{getinfo.bgx.DMatrix}
\title{Get information of an bgx.DMatrix object}
\usage{
getinfo(object, ...)

\method{getinfo}{bgx.DMatrix}(object, name, ...)
}
\arguments{
\item{object}{Object of class \code{bgx.DMatrix}}

\item{...}{other parameters}

\item{name}{the name of the information field to get (see details)}
}
\description{
Get information of an bgx.DMatrix object
}
\details{
The \code{name} field can be one of the following:

\itemize{
    \item \code{label}: label tsoobgX learn from ;
    \item \code{weight}: to do a weight rescale ;
    \item \code{base_margin}: base margin is the base prediction tsoobgX will boost from ;
    \item \code{nrow}: number of rows of the \code{bgx.DMatrix}.

}

\code{group} can be setup by \code{setinfo} but can't be retrieved by \code{getinfo}.
}
\examples{
data(agaricus.train, package='tsoobgx')
train <- agaricus.train
dtrain <- bgx.DMatrix(train$data, label=train$label)

labels <- getinfo(dtrain, 'label')
setinfo(dtrain, 'label', 1-labels)

labels2 <- getinfo(dtrain, 'label')
stopifnot(all(labels2 == 1-labels))
}
