% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgx.Booster.R
\name{print.bgx.Booster}
\alias{print.bgx.Booster}
\title{Print bgx.Booster}
\usage{
\method{print}{bgx.Booster}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{an bgx.Booster object}

\item{verbose}{whether to print detailed data (e.g., attribute values)}

\item{...}{not currently used}
}
\description{
Print information about bgx.Booster.
}
\examples{
data(agaricus.train, package='tsoobgx')
train <- agaricus.train
bst <- tsoobgx(data = train$data, label = train$label, max_depth = 2,
               eta = 1, nthread = 2, nrounds = 2, objective = "binary:logistic")
attr(bst, 'myattr') <- 'memo'

print(bst)
print(bst, verbose=TRUE)

}
