% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgx.DMatrix.R
\name{slice}
\alias{slice}
\alias{slice.bgx.DMatrix}
\alias{[.bgx.DMatrix}
\title{Get a new DMatrix containing the specified rows of
original bgx.DMatrix object}
\usage{
slice(object, ...)

\method{slice}{bgx.DMatrix}(object, idxset, ...)

\method{[}{bgx.DMatrix}(object, idxset, colset = NULL)
}
\arguments{
\item{object}{Object of class "bgx.DMatrix"}

\item{...}{other parameters (currently not used)}

\item{idxset}{a integer vector of indices of rows needed}

\item{colset}{currently not used (columns subsetting is not available)}
}
\description{
Get a new DMatrix containing the specified rows of
original bgx.DMatrix object
}
\examples{
data(agaricus.train, package='tsoobgx')
train <- agaricus.train
dtrain <- bgx.DMatrix(train$data, label=train$label)

dsub <- slice(dtrain, 1:42)
labels1 <- getinfo(dsub, 'label')
dsub <- dtrain[1:42, ]
labels2 <- getinfo(dsub, 'label')
all.equal(labels1, labels2)

}
